Changed the input type to int - this should be more efficient and more clear that the input type is ranks and not some other measure. Also fixed a small bug in the python ranking code

+++ python/IDR_parameter_estimation.h
@@ -25,8 +25,8 @@ struct OptimizationRV
 struct OptimizationRV
 em_gaussian(
     int n_samples,
-    double* x, 
-    double* y,
+    int* x, 
+    int* y,
     double* IDRs,
     int print_status_msgs);
